# openapi.yaml
openapi: 3.0.3
info:
  title: Last Price
  description: |
    Server for conducting A/B testing experiments on pricing strategies using Last Price integration.
    
    This server provides endpoints for:
    - Assigning users to experiment variants
    - Tracking pricing page views and conversions
    - Processing Paid.ai webhook events
    - Analyzing experiment results
    
    **Architecture**: The system uses deterministic SHA-256 hashing to ensure consistent user-to-variant mapping,
    emits signals to Paid.ai's Signals API for event tracking, and provides real-time analytics.
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /contacts:
    get:
      tags:
        - Contacts
      summary: List all contacts
      description: Retrieves all contacts from Last Price
      operationId: listLastPriceContacts
      responses:
        '200':
          description: Retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LastPriceContact'

    post:
      tags:
        - Contacts
      summary: Create a contact
      description: Creates a new contact in Last Price
      operationId: createLastPriceContact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LastPriceContactInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LastPriceContact'
...
  /contacts/external/{externalId}:
    delete:
      tags:
        - Contacts
      summary: Delete contact by external ID
      description: Deletes a contact by external ID from Last Price
      operationId: deleteLastPriceContactByExternalId
      parameters:
        - name: externalId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
  /agents:
    get:
      tags:
        - Agents
      summary: List all agents
      description: Retrieves all agents from Last Price
      operationId: listLastPriceAgents
      responses:
        '200':
          description: Retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LastPriceAgent'

    post:
      tags:
        - Agents
      summary: Create an agent
      description: Creates a new agent in Last Price
      operationId: createLastPriceAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LastPriceAgentInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LastPriceAgent'
  /agents/{agentId}:
    get:
      tags:
        - Agents
      summary: Retrieve an agent
      description: Retrieves an agent by ID from Last Price
      operationId: getLastPriceAgent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LastPriceAgent'

    put:
      tags:
        - Agents
      summary: Update an agent
      description: Updates an agent in Last Price
      operationId: updateLastPriceAgent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LastPriceAgentInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LastPriceAgent'

    delete:
      tags:
        - Agents
      summary: Delete an agent
      description: Deletes an agent from Last Price
      operationId: deleteLastPriceAgent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
  /agents/external/{externalId}:
    get:
      tags:
        - Agents
      summary: Retrieve agent by external ID
      description: Retrieves an agent by external ID from Last Price
      operationId: getLastPriceAgentByExternalId
      parameters:
        - name: externalId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LastPriceAgent'

    put:
      tags:
        - Agents
      summary: Update agent by external ID
      description: Updates an agent by external ID in Last Price
      operationId: updateLastPriceAgentByExternalId
      parameters:
        - name: externalId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LastPriceAgentInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LastPriceAgent'

    delete:
      tags:
        - Agents
      summary: Delete agent by external ID
      description: Deletes an agent by external ID from Last Price
      operationId: deleteLastPriceAgentByExternalId
      parameters:
        - name: externalId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
...
  /costs/analytics:
    get:
      tags:
        - Costs
      summary: Get cost analytics
      description: Retrieves aggregated cost analytics
      operationId: getLastPriceCostAnalytics
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [agent, customer, day, week, month]
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LastPriceCostAnalytics'
components:
  schemas:
...
    LastPriceContact:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        accountId:
          type: string
          format: uuid
        salutation:
          type: string
          enum: ["Mr.", "Mrs.", "Miss.", "Ms.", "Dr.", "Prof."]
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        billingStreet:
          type: string
        billingCity:
          type: string
        billingStateProvince:
          type: string
        billingCountry:
          type: string
        billingZipPostalCode:
          type: string
        externalId:
          type: string
        accountName:
          type: string
          readOnly: true
    LastPriceContactInput:
      type: object
      required:
        - firstName
        - lastName
        - email
      properties:
        accountId:
          type: string
          format: uuid
        salutation:
          type: string
          enum: ["Mr.", "Mrs.", "Miss.", "Ms.", "Dr.", "Prof."]
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        billingStreet:
          type: string
        billingCity:
          type: string
        billingStateProvince:
          type: string
        billingCountry:
          type: string
        billingZipPostalCode:
          type: string
        externalId:
          type: string
    LastPricePriceTier:
      type: object
      properties:
        minQuantity:
          type: integer
        maxQuantity:
          type: integer
        unitPrice:
          type: number
    LastPricePricePoint:
      type: object
      properties:
        unitPrice:
          type: number
        minQuantity:
          type: integer
        includedQuantity:
          type: integer
        tiers:
          type: array
          items:
            $ref: '#/components/schemas/LastPricePriceTier'
    LastPricePricing:
      type: object
      properties:
        eventName:
          type: string
        chargeType:
          type: string
          enum: [one_time, recurring, usage]
        pricingModel:
          type: string
          enum: [flat_fee, per_unit, tiered_pricing, graduated_pricing, volume_pricing]
        billingFrequency:
          type: string
          enum: [monthly, quarterly, annually]
        taxable:
          type: boolean
        pricePoints:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/LastPricePricePoint'
    LastPriceAgentAttribute:
      type: object
      properties:
        name:
          type: string
        active:
          type: boolean
        pricing:
          $ref: '#/components/schemas/LastPricePricing'
    LastPriceAgent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        active:
          type: boolean
        agentCode:
          type: string
        externalId:
          type: string
        agentAttributes:
          type: array
          items:
            $ref: '#/components/schemas/LastPriceAgentAttribute'
    LastPriceAgentInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
        active:
          type: boolean
        agentCode:
          type: string
        externalId:
          type: string
        agentAttributes:
          type: array
          items:
            $ref: '#/components/schemas/LastPriceAgentAttribute'
...
    LastPriceCostAnalytics:
      type: object
      properties:
        totalCost:
          type: number
        currency:
          type: string
        period:
          type: object
          properties:
            startDate:
              type: string
              format: date
            endDate:
              type: string
              format: date
        breakdown:
          type: array
          items:
            type: object
            properties:
              groupBy:
                type: string
              key:
                type: string
              cost:
                type: number
              percentage:
                type: number
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token for Last Price API authentication
    webhookSignature:
      type: apiKey
      in: header
      name: x-paid-signature
      description: HMAC signature for webhook verification